move per-stream setup into start()
