+++ airbyte-integrations/connectors/destination-s3-data-lake/src/main/kotlin/io/airbyte/integrations/destination/s3_data_lake/S3DataLakeStreamLoader.kt
@@ -6,7 +6,7 @@ package io.airbyte.integrations.destination.s3_data_lake
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings
 import io.airbyte.cdk.load.command.DestinationStream
-import io.airbyte.cdk.load.data.MapperPipeline
+import io.airbyte.cdk.load.data.iceberg.parquet.IcebergParquetPipelineFactory
 import io.airbyte.cdk.load.message.Batch
 import io.airbyte.cdk.load.message.DestinationRecordAirbyteValue
 import io.airbyte.cdk.load.message.SimpleBatch
@@ -18,17 +18,46 @@ import io.airbyte.integrations.destination.s3_data_lake.io.S3DataLakeUtil
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.apache.iceberg.Table
 
+private val logger = KotlinLogging.logger {}
+
 @SuppressFBWarnings("NP_NONNULL_PARAM_VIOLATION", justification = "Kotlin async continuation")
 class S3DataLakeStreamLoader(
+    private val icebergConfiguration: S3DataLakeConfiguration,
     override val stream: DestinationStream,
-    private val table: Table,
+    private val s3DataLakeTableSynchronizer: S3DataLakeTableSynchronizer,
     private val s3DataLakeTableWriterFactory: S3DataLakeTableWriterFactory,
     private val s3DataLakeUtil: S3DataLakeUtil,
-    private val pipeline: MapperPipeline,
     private val stagingBranchName: String,
     private val mainBranchName: String
 ) : StreamLoader {
-    private val log = KotlinLogging.logger {}
+    private lateinit var table: Table
+    private val pipeline = IcebergParquetPipelineFactory().create(stream)
+
+    override suspend fun start() {
+        val properties = s3DataLakeUtil.toCatalogProperties(config = icebergConfiguration)
+        val catalog = s3DataLakeUtil.createCatalog(DEFAULT_CATALOG_NAME, properties)
+        val incomingSchema = s3DataLakeUtil.toIcebergSchema(stream = stream, pipeline = pipeline)
+        table =
+            s3DataLakeUtil.createTable(
+                streamDescriptor = stream.descriptor,
+                catalog = catalog,
+                schema = incomingSchema,
+                properties = properties
+            )
+
+        s3DataLakeTableSynchronizer.applySchemaChanges(table, incomingSchema)
+
+        try {
+            logger.info {
+                "maybe creating branch $DEFAULT_STAGING_BRANCH for stream ${stream.descriptor}"
+            }
+            table.manageSnapshots().createBranch(DEFAULT_STAGING_BRANCH).commit()
+        } catch (e: IllegalArgumentException) {
+            logger.info {
+                "branch $DEFAULT_STAGING_BRANCH already exists for stream ${stream.descriptor}"
+            }
+        }
+    }
 
     override suspend fun processRecords(
         records: Iterator<DestinationRecordAirbyteValue>,
@@ -42,7 +71,7 @@ class S3DataLakeStreamLoader(
                 importType = stream.importType
             )
             .use { writer ->
-                log.info { "Writing records to branch $stagingBranchName" }
+                logger.info { "Writing records to branch $stagingBranchName" }
                 records.forEach { record ->
                     val icebergRecord =
                         s3DataLakeUtil.toRecord(
@@ -64,7 +93,7 @@ class S3DataLakeStreamLoader(
                     writeResult.dataFiles().forEach { append.appendFile(it) }
                     append.commit()
                 }
-                log.info { "Finished writing records to $stagingBranchName" }
+                logger.info { "Finished writing records to $stagingBranchName" }
             }
 
         return SimpleBatch(Batch.State.COMPLETE)
@@ -74,12 +103,12 @@ class S3DataLakeStreamLoader(
         if (streamFailure == null) {
             // Doing it first to make sure that data coming in the current batch is written to the
             // main branch
-            log.info {
+            logger.info {
                 "No stream failure detected. Committing changes from staging branch '$stagingBranchName' to main branch '$mainBranchName."
             }
             table.manageSnapshots().fastForwardBranch(mainBranchName, stagingBranchName).commit()
             if (stream.minimumGenerationId > 0) {
-                log.info {
+                logger.info {
                     "Detected a minimum generation ID (${stream.minimumGenerationId}). Preparing to delete obsolete generation IDs."
                 }
                 val generationIdsToDelete =
@@ -93,7 +122,7 @@ class S3DataLakeStreamLoader(
                     generationIdsToDelete
                 )
                 //  Doing it again to push the deletes from the staging to main branch
-                log.info {
+                logger.info {
                     "Deleted obsolete generation IDs up to ${stream.minimumGenerationId - 1}. " +
                         "Pushing these updates to the '$mainBranchName' branch."
                 }
